// frontend/certifications-ui/src/app/core/api/generated/certificationsApiV1.schemas.ts
export interface CertificationDto {
  certificationId: number;
  contractId: number;
  assessor: string;
  certificationDate: string;
  /** @nullable */
  protocolDate?: string | null;
  /** @nullable */
  prolongationSend?: string | null;
  /** @nullable */
  prolongationReturned?: string | null;
  isCompleted: boolean;
}

export type CertificationOverviewRowDtoStatus =
  (typeof CertificationOverviewRowDtoStatus)[keyof typeof CertificationOverviewRowDtoStatus];

export const CertificationOverviewRowDtoStatus = {
  NotApplicable: 'NotApplicable',
  ContractValid: 'ContractValid',
  CertificationPending: 'CertificationPending',
  CertificationInProgress: 'CertificationInProgress',
  CertificationMissing: 'CertificationMissing',
} as const;

export interface CertificationOverviewRowDto {
  employeeId: string;
  personalId: string;
  firstName: string;
  /** @nullable */
  middleName?: string | null;
  lastName: string;
  isAdmin: boolean;
  /** @nullable */
  contractId?: number | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  division?: string | null;
  /** @nullable */
  contractDate?: string | null;
  /** @nullable */
  effectiveValidTo?: string | null;
  status: CertificationOverviewRowDtoStatus;
  latestCertification?: CertificationDto | null;
}

export interface CertificationOverviewRowDtoPagedResult {
  items: CertificationOverviewRowDto[];
  page: number;
  pageSize: number;
  totalCount: number;
}
// ...
export type GetCertificationOverviewParams = {
  page?: number;
  pageSize?: number;
  name?: string;
  department?: string;
  /**
   * @nullable
   */
  status?: GetCertificationOverviewStatus;
  validToFrom?: string;
  validToTo?: string;
  includeInactive?: boolean;
  sort?: string;
  direction?: string;
};

export type GetCertificationOverviewStatus =
  (typeof GetCertificationOverviewStatus)[keyof typeof GetCertificationOverviewStatus] | null;

export const GetCertificationOverviewStatus = {
  NotApplicable: 'NotApplicable',
  ContractValid: 'ContractValid',
  CertificationPending: 'CertificationPending',
  CertificationInProgress: 'CertificationInProgress',
  CertificationMissing: 'CertificationMissing',
} as const;